+++ generator/src/Factory/Pdf.hs
@@ -30,7 +30,7 @@ import Data.Text (Text)
 import Factory.Domain
 import Factory.Geometry (rectangleToBoard)
 import Factory.Interpreter (ColorSpaceResource (..), Resources (Resources), VisualResource (..), interpretOperators)
-import Factory.Vectorize (ArtworkPartition (..), ImageDisposition (..), classifyImage, opaqueHighlighter, partitionArtwork, traceImage)
+import Factory.Vectorize (ArtworkPartition (..), ImageDisposition (..), classifyImage, opaqueHighlighter, partitionArtwork, traceImage, traceSmoothImage)
 import Network.URI (URI (uriAuthority, uriFragment, uriPath, uriQuery, uriScheme), URIAuth (uriPort, uriRegName, uriUserInfo), parseURI)
 import Pdf.Content (Expr, Operator, parseContent, readNextOperator)
 import Pdf.Core
@@ -236,10 +236,15 @@ prepareImage pdf resourceName reference stream dictionary = do
     unsupported -> throwError (UnsupportedImage ("unsupported image filter: " <> unsupported))
 
 prepareArtwork :: Asset -> ArtworkPartition -> PdfAction PreparedImage
-prepareArtwork asset partition = case partition of
-  TraceableArtwork image -> PreparedVector <$> liftEither (traceImage image)
-  UntraceableArtwork image -> pure (PreparedPng asset image)
-  MixedArtwork traceable residual -> (\shapes -> PreparedMixed shapes asset residual) <$> liftEither (traceImage traceable)
+prepareArtwork asset partition = do
+  traced <- traverse (liftEither . traceImage) (tracedComponents partition)
+  smoothed <- traverse (liftEither . traceSmoothImage) (smoothedComponents partition)
+  let shapes = concat (catMaybes [traced, smoothed])
+  pure $ case residualComponents partition of
+    Nothing -> PreparedVector shapes
+    Just residual
+      | null shapes -> PreparedPng asset residual
+      | otherwise -> PreparedMixed shapes asset residual
 
 materializeImages :: FilePath -> [PreparedImage] -> PdfAction [Asset]
 materializeImages assetDirectory preparedImages = do